Fix streaming
In the interest of full disclosure, I did not test this on older versions of aiohttp and will assume it used to work at some point. I am currently on aiohttp v `3.7.4.post0` and the changes I've made produce a working stream.

---

Since `FirebaseHTTP._request` puts `".json"` on the url, it makes sense to also do it in the stream, where it is absolutely needed (if it's not already provided...)

Besides fixing the functionality of the method itself, the additional space character on line 76 is because the event would otherwise be `" put"` for example.

Technically this is a breaking change, but if you're up to date on aiohttp, it was already broken. 🤪